Delete ignored tests - We have one wmsc parse test left, also rename WmsC to Wmsc in tests
