+++ Tests/BruTile.Tests/Resources/Wmsc/WmscCapabilities_1_1_1_.xml
